+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.13.3] 2025-04-29
 
 ### Added
 
+- Added new menu commands provided by [@sttk3](https://community.adobe.com/t5/illustrator-discussions/executemenucommand-command-list/m-p/15377798/page/2#M445281)
+    - Object > Generative Expand... > Make...
+    - Object > Generative Expand... > Combine
+    - Object > Print Bleed...
 - Logger for debugging (with accompanying Enable/Disable Logging command)
 
 ### Fixed